+++ CHANGELOG.md
@@ -10,5 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Automatic creation of test databases and dropping them upon test completion
  - Support for loading fixture sql files
  - Support for loading table data into already existing tables
+ - Support for loading CSV file into existing table
+ - Support for fetching table data from the database as an array
 
 [Unreleased]: https://github.com/EcomDev/mysql-test-utils/compare/67dda02cdfe2a2ca766b5821da2f48b2c8a1ef68...develop